requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# HTTP Client
httpx>=0.24.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment Variables
python-dotenv==1.0.0

# ZeroDB SDK
zerodb-mcp==1.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-httpx==0.27.0

# Development Tools
black==24.1.1
flake8==7.0.0
mypy==1.8.0
